RFC: Minimum viable way to keep the partition status cache fast when it is up to date, even during in progress runs
Summary:
This is a proof of concept for the simplest possible way to keep the asset status partition cache from needing to do any additional work when it is up to date, once we have an ASSET_MATERIALIZATION_FAILURE to work with. It is much less significant of a change than I had been originally imagining I would make - all it does is use the presence of a new FAILURE event to indicate that it should update the in progress and failed statuses. If there has been no new event since the last time it updated, it just uses the existing values. This should make the cache fast to compute when it is up to date, even when there are in progress runs.

What this does not help with at all yet is making the process of rebuilding it fast when an event does come and and it needs to update the cache - it still needs to start with the earliest in progress run, recompute the in progress subset from scratch, etc.